refactor(cli): move cdktf-project into separate file
